Refactoring finished. Read all events from a stream forward working
