+++ lib/estore.rb
@@ -7,6 +7,4 @@ require 'estore/connection_context'
 require 'estore/connection'
 require 'estore/connection/buffer'
 require 'estore/connection/protocol'
-require 'estore/subscription'
-require 'estore/catchup_subscription'
 require 'estore/commands'
