+++ lib/estore/commands/ping.rb
@@ -1,0 +1,16 @@
+module Estore
+  module Commands
+    class Ping
+      include Command
+
+      def call
+        write('Ping')
+        promise
+      end
+
+      def handle(*)
+        promise.fulfill('Pong')
+      end
+    end
+  end
+end
